ugly fix overflow in hate and damage AggroInfo values
git-svn-id: http://svn.l2jserver.com/trunk@4171 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/com/l2jserver/gameserver/model/actor/L2Attackable.java
@@ -113,7 +113,7 @@ public class L2Attackable extends L2Npc
 
 		public final void addHate(int value)
 		{
-			_hate = Math.min(_hate + value, 999999999);
+			_hate = (int)Math.min(_hate + (long)value, 999999999);
 		}
 
 		public final void stopHate()
@@ -128,7 +128,7 @@ public class L2Attackable extends L2Npc
 
 		public final void addDamage(int value)
 		{
-			_damage = Math.min(_damage + value, 999999999);
+			_damage = (int)Math.min(_damage + (long)value, 999999999);
 		}
 
 		@Override
